+++ scrapedia/errors.py
@@ -1,11 +1,30 @@
-# -*- coding: utf-8 -*-
+"""Scrapedia's source of exception classes.
 
-class FutpediaRequestError(Exception):
-	"""To be raised when an HTTP request error occurs."""
+Classes: ScrapediaError, ScrapediaRequestError, ScrapediaParseError,
+ScrapediaNotFoundError.
+"""
+
+class ScrapediaError(Exception):
+	"""Generic error to be implemented by further classes in order to uncouple
+	Scrapedia's exceptions from others.
+	"""
 	pass
 
-class FutpediaNotFoundError(Exception):
-	"""To be raised when an expected HTML content from a page is missing or
-	when a chosen team or championship is not found.
+
+class ScrapediaRequestError(ScrapediaError):
+	"""To be raised whenever an error related to an HTTP request of a Futpédia
+	page occurs.
 	"""
+	pass
+
+
+class ScrapediaParseError(ScrapediaError):
+	"""To be raised whenever an error related to parsing occurs like missing
+	expected content from a page or data conversion failures.
+	"""
+	pass
+
+
+class ScrapediaNotFoundError(ScrapediaError):
+	"""To be raised whenever there is an error where a chosen team, game or championship metadata is not found on the requested or cached data."""
 	pass